+++ src/main/java/znu/visum/components/diary/usecases/query/application/types/DiaryYear.java
@@ -1,17 +1,16 @@
 package znu.visum.components.diary.usecases.query.application.types;
 
 import lombok.Getter;
-import znu.visum.components.diary.domain.DiaryMovie;
+import znu.visum.components.diary.domain.Diary;
 import znu.visum.core.models.common.Month;
 
-import java.time.Year;
-import java.util.*;
+import java.util.List;
 import java.util.stream.Collectors;
 
 @Getter
 public class DiaryYear {
 
-  private static final List<Month> MONTH_VALUES =
+  private static final List<Month> MONTHS =
       List.of(
           Month.DECEMBER,
           Month.NOVEMBER,
@@ -28,61 +27,19 @@ public class DiaryYear {
   private final int year;
   private final List<DiaryMonth> months;
 
-  public DiaryYear(int year, List<DiaryMonth> months) {
+  private DiaryYear(int year, List<DiaryMonth> months) {
     this.months = months;
     this.year = year;
   }
 
-  public static DiaryYear from(Year year, List<DiaryMovie> movies) {
-    int yearValue = year.getValue();
+  public static DiaryYear from(Diary diary) {
+    int yearValue = diary.getYear().getValue();
 
-    Map<znu.visum.core.models.common.Month, List<DiaryDay>> byMonth = new HashMap<>();
+    List<DiaryMonth> months =
+        MONTHS.stream()
+            .map(month -> DiaryMonth.from(diary.getMonths().get(Month.toJava(month))))
+            .collect(Collectors.toUnmodifiableList());
 
-    for (DiaryMovie movie : movies) {
-      DiaryDay dayResponse =
-          new DiaryDay(movie.getViewingDate().getDayOfMonth(), DiaryMovieType.from(movie));
-
-      znu.visum.core.models.common.Month month = Month.from(movie.getViewingDate().getMonth());
-      if (byMonth.containsKey(month)) {
-        List<DiaryDay> monthMovies = byMonth.get(month);
-        monthMovies.add(dayResponse);
-      } else {
-        byMonth.put(month, new ArrayList<>(Arrays.asList(dayResponse)));
-      }
-    }
-
-    return new DiaryYear(
-        yearValue,
-        MONTH_VALUES.stream()
-            .map(
-                month -> {
-                  // Sort by day
-                  if (byMonth.get(month) != null) {
-                    Comparator<DiaryDay> byDay = Comparator.comparing(DiaryDay::getDay);
-                    byMonth.get(month).sort(byDay.reversed());
-                  }
-
-                  return new DiaryMonth(
-                      month, byMonth.get(month) != null ? byMonth.get(month) : new ArrayList<>());
-                })
-            .collect(Collectors.toList()));
-  }
-
-  @Override
-  public String toString() {
-    return "DiaryYear{" + "months='" + months + "'," + "year='" + year + "'" + "}";
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    DiaryYear that = (DiaryYear) o;
-    return java.util.Objects.equals(months, that.months) && year == that.year;
-  }
-
-  @Override
-  public int hashCode() {
-    return java.util.Objects.hash(months, year);
+    return new DiaryYear(yearValue, months);
   }
 }