+++ src/main/java/znu/visum/components/diary/usecases/query/domain/GetDiaryByYear.java
@@ -2,17 +2,16 @@ package znu.visum.components.diary.usecases.query.domain;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import znu.visum.components.diary.domain.DiaryMovie;
-import znu.visum.components.history.domain.ViewingEntry;
-import znu.visum.components.movies.domain.Movie;
+import znu.visum.components.diary.domain.Diary;
+import znu.visum.components.movies.domain.DiaryFilters;
+import znu.visum.components.movies.domain.MovieDiaryFragment;
 import znu.visum.components.movies.domain.MovieQueryRepository;
 
-import java.time.Year;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class GetDiaryByYear {
+
   private final MovieQueryRepository movieQueryRepository;
 
   @Autowired
@@ -20,24 +19,10 @@ public class GetDiaryByYear {
     this.movieQueryRepository = movieQueryRepository;
   }
 
-  public List<DiaryMovie> getDiaryMovies(Year year, Integer grade, Long genreId) {
-    int yearValue = year.getValue();
+  public Diary getDiary(DiaryFilters filters) {
+    List<MovieDiaryFragment> moviesSeenDuringTheYear =
+        this.movieQueryRepository.findByDiaryFilters(filters);
 
-    List<Movie> moviesSeenDuringTheYear =
-        this.movieQueryRepository.findByDiaryFilters(year, grade, genreId);
-
-    List<DiaryMovie> diaryMovies = new ArrayList<>();
-
-    for (Movie movie : moviesSeenDuringTheYear) {
-      for (ViewingEntry viewingEntry : movie.getViewingHistory().getEntries()) {
-        if (viewingEntry.getDate() == null || viewingEntry.getDate().getYear() != yearValue) {
-          continue;
-        }
-
-        diaryMovies.add(DiaryMovie.from(movie, viewingEntry.getId()));
-      }
-    }
-
-    return diaryMovies;
+    return new Diary(filters.getYear(), moviesSeenDuringTheYear);
   }
 }